# artifacts/APISpecificationOPENAPI.yaml
openapi: 3.0.3
info:
  title: SLPVT API
  description: Sri Lanka Police Vehicle Tracking System — REST API for managing stations, vehicles, drivers, tracking devices, and GPS location data.
  version: 1.0.0
  contact:
    name: SLPVT Development Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from POST /auth/login
    DeviceApiKey:
      type: apiKey
      in: header
      name: X-Device-Key
      description: Raw API key issued to a tracking device at provisioning time
  parameters:
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    sortByParam:
      name: sortBy
      in: query
      schema:
        type: string
    orderParam:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        description:
          type: string
          nullable: true
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                nullable: true
              message:
                type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 401
            message: Authorization header missing or malformed
            description: null
            errors: []
    Forbidden:
      description: Authenticated but insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 403
            message: Role STATION_OFFICER does not have permission to perform this action
            description: null
            errors: []
...
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 409
            message: A station with this name already exists
            description: null
            errors: []
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 422
            message: Validation failed
            description: One or more fields in the request are invalid
            errors:
              - field: name
                message: name is required
...
paths:
...
  /vehicles:
    get:
      tags: [Vehicles]
      summary: List vehicles
      description: Returns a paginated list of active vehicles with optional filters. Scope is enforced based on the requesting officer's jurisdiction.
      security:
        - BearerAuth: []
      parameters:
        - name: registrationNumber
          in: query
          schema:
            type: string
        - name: ownerNic
          in: query
          schema:
            type: string
        - name: ownerName
          in: query
          schema:
            type: string
        - name: policeStatus
          in: query
          schema:
            type: string
            enum: [CLEAN, STOLEN, WANTED, SUSPENDED]
        - name: makeModel
          in: query
          schema:
            type: string
        - name: hasDevice
          in: query
          schema:
            type: boolean
        - name: dsDivisionId
          in: query
          schema:
            type: string
            format: uuid
        - name: districtId
          in: query
          schema:
            type: string
            format: uuid
        - name: provinceId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/sortByParam'
        - $ref: '#/components/parameters/orderParam'
      responses:
        "200":
          description: Paginated list of vehicles
          content:
            application/json:
              example:
                count: 205
                offset: 0
                limit: 20
                next: /api/v1/vehicles?offset=20&limit=20
                previous: null
                data:
                  - vehicle_id: 991e8400-e29b-41d4-a716-446655440000
                    vehicle_reference_id: REF00001
                    registration_number: WP AAB-1234
                    chassis_number: CHASSIS00000001
                    color: Yellow
                    make_model: Bajaj RE
                    police_status: CLEAN
                    owner_full_name: Nuwan Perera
                    owner_contact: "+94771234567"
                    device_serial_number: SN-TRK-00001
                    device_admin_status: ACTIVE
                    ds_division_name: Colombo
                    district_name: Colombo
                    province_name: Western
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Vehicles]
      summary: Register a new vehicle
      description: Registers a tuk-tuk imported from the DMT database.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            example:
              vehicle_reference_id: REF00206
              registration_number: WP ZZZ-9999
              chassis_number: CHASSIS00000206
              color: Blue
              make_model: TVS King
              owner_nic: "199012345678"
              owner_full_name: Kamal Silva
              owner_contact: "+94772345678"
              ds_division_id: some-uuid
      responses:
        "201":
          description: Vehicle registered
          content:
            application/json:
              example:
                data:
                  vehicle_id: aa2e9511-f30c-52e5-b827-557766553333
                  vehicle_reference_id: REF00206
                  registration_number: WP ZZZ-9999
                  chassis_number: CHASSIS00000206
                  color: Blue
                  make_model: TVS King
                  police_status: CLEAN
                  owner_full_name: Kamal Silva
                  owner_contact: "+94772345678"
                  ds_division_id: some-uuid
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          $ref: '#/components/responses/Conflict'
        "422":
          $ref: '#/components/responses/ValidationError'
...